Refactoring
apx_nodeData has been split into 3 new classes (apx_nodeInfo,
apx_nodeData2, apx_nodeInstance)

Removed old event handler located inside apx_fileManager

Implement APX client based on new file and node managers

Major refactoring of file manager and node manager

Refactor how transmithandler is initialized in connections

Refactor how fileInfo is transmitted and processed
Cleaned up old files (From old apx stack)

Updated adt and cutil to latest

+++ apx/common/test/apx_transmitHandlerSpy.c
@@ -101,7 +101,7 @@ uint8_t* apx_transmitHandlerSpy_getSendBuffer(void *arg, int32_t msgLen)
 int32_t apx_transmitHandlerSpy_send(void *arg, int32_t offset, int32_t msgLen)
 {
    apx_transmitHandlerSpy_t* self = (apx_transmitHandlerSpy_t*) arg;
-   if ( (self != 0) && (adt_bytearray_length(self->buf) >= msgLen) )
+   if ( (self != 0) && (adt_bytearray_length(self->buf) >= (uint32_t) msgLen) )
    {
       adt_ary_push(self->transmitted, self->buf);
       self->buf = 0;
